move batchUtils and dateUtils to lib
