+++ .env
@@ -1,2 +1,2 @@
-VITE_BASE_URL=https://ims-backend-app.onrender.com
-#VITE_BASE_URL=http://localhost:8000
+#VITE_BASE_URL=https://ims-backend-app.onrender.com
+VITE_BASE_URL=http://localhost:8000
