+++ src/lib/batchCardUtils.jsx
